Корректировка структуры проекта
