+++ src/main/java/ru/practicum/shareit/item/controller/ItemController.java
@@ -1,12 +1,12 @@
-package ru.practicum.shareit.booking;
+package ru.practicum.shareit.item.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
- * TODO Sprint add-bookings.
+ * TODO Sprint add-controllers.
  */
 @RestController
-@RequestMapping(path = "/bookings")
-public class BookingController {
+@RequestMapping("/items")
+public class ItemController {
 }
